ported changes from ZDoom revision 3883-4001 and from GZDoom revision 1488
* Updated to ZDoom 4001:
- Fixed: When the player sprite is being drawn with a fixed colormap, hardware acceleration should not be disabled when the regular colormap has lights that stay bright throughout.
- Fixed: The originx and originy for the textured automap is supposed to be calculated after rotating originpt by the automap, not before.
- Fixed: Sky floors should not be drawn in the textured automap.
- Fixed: The Fist had the wrong Tag string.
- Changed default FluidSynth reverb and chorus parameters to the ones suggested here: http://forums.scummvm.org/viewtopic.php?p=72972&sid=d139e99a13359541ead07073112e3888#72972
- Fixed: Changing fluid_chorus and fluid_reverb did not take effect until the next song started.
- Changed synth.chorus.active and synth.reverb.active settings to use fluid_settings_setint(), since they're int and not string settings.
- Fixed: FluidSynthMIDIDevice::FluidSettingInt() interpreted the return value from fluid_settings_setint() wrongly.
- Fix type conversion warning in file_wad.cpp.
- Do not crash when reading ATAG chunks on processors that don't like unaligned accesses.
- Added support for Eternity's JUMP table and PCD_GOTOSTACK (again, requires advanced compiler).
- Added support for Eternity's ATAG chunk (mixed-type arrays).  Requires a structure aware compiler (not ACC).

git-svn-id: http://mancubus.net/svn/hosted/gzdoom/trunk@1488 b0f79afe-0144-0410-b225-9a4edf0717df

+++ src/p_acs.cpp
@@ -7275,7 +7275,7 @@ scriptwait:
 			if (activationline != NULL)
 			{
 				activationline->special = 0;
-				DPrintf("Cleared line special on line %d\n", activationline - lines);
+				DPrintf("Cleared line special on line %d\n", (int)(activationline - lines));
 			}
 			break;
